news_summarizer/requirements.txt
fastapi
uvicorn[standard]
feedparser
trafilatura
httpx
sentence_transformers
scikit-learn
python-dotenv